+++ accounts/forms.py
@@ -1,0 +1,9 @@
+from django.contrib.auth.forms import UserCreationForm
+
+from .models import User
+
+
+class SignUpForm(UserCreationForm):
+    class Meta:
+        model = User
+        fields = ["email", "username"]
