+++ accounts/templates/accounts/index.html
@@ -3,5 +3,5 @@
 
 {% block content %} 
 <h1>トップページ</h1>
-<p><a href="{% url 'accounts:signup' %}">ユーザ作成</a></p>
+<p><a href="{% url 'kakeiboapp:signup' %}">ユーザ作成</a></p>
 {% endblock %}
