+++ accounts/templates/accounts/signup.html
@@ -1,0 +1,22 @@
+{% extends "kakeiboapp/base.html" %}
+{% block title %}ユーザ登録{% endblock %}
+
+{% block content %}
+<div class="row">
+    <div class="col-sm-6">
+        <br>
+        <h1>ユーザ登録</h1>
+        <form method="POST">
+            {% csrf_token %}
+            {{ form.non_field_errors }}
+            {% for field in form %}
+                {{ field.label }}
+                {{ field }}
+                {{ field.errors }}
+                <br>
+            {% endfor %}
+            <button type="submit">登録</button>
+        </form>
+    </div>
+</div>
+{% endblock %}
