+++ kakeiboapp/forms.py
@@ -1,9 +1,2 @@
-from django.contrib.auth.forms import UserCreationForm
+
 # from django.contrib.auth.models import User
-from .accounts.models import User
-
-
-class SignUpForm(UserCreationForm):
-    class Meta:
-        model = User
-        fields = ["username", "email", "password1", "password2"]
